+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/adminamendlocationdetails/AdminAmendLocationDetailsAboutToStartTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit;
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.adminamendlocationdetails;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
